Ставка должна быть float в 2 ПР2.3

+++ ПР2.3/task2.py
@@ -3,7 +3,7 @@
 
 
 class Worker:
-    def __init__(self, name: str, surname: str, rate: int, days: int):
+    def __init__(self, name: str, surname: str, rate: float, days: int):
         self.__name = name
         self.__surname = surname
         self.__rate = rate
